+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog also contains important changes in dependencies.
 ## [Unreleased]
 ### Fixed
 - `fill-opacity` and `stroke-opacity` resolving.
+- Double `transform` when resolving `symbol`.
 
 ## [0.30.0] - 2023-03-25
 ### Added